+++ src/devices/ros2RGBDConversionUtils/CMakeLists.txt
@@ -8,20 +8,21 @@
 #  return()
 #endif()
 
-add_library(Ros2ConversionUtils OBJECT)
+add_library(Ros2RGBDConversionUtils OBJECT)
 
-target_sources(Ros2ConversionUtils PRIVATE Ros2ConversionUtils.cpp
-                                              Ros2ConversionUtils.h
-                                              ros2PixelCode.h
-                                              ros2PixelCode.cpp)
+target_sources(Ros2RGBDConversionUtils PRIVATE
+        Ros2RGBDConversionUtils.cpp
+        Ros2RGBDConversionUtils.h
+        ros2PixelCode.h
+        ros2PixelCode.cpp)
                                               
-target_include_directories(Ros2ConversionUtils PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_include_directories(Ros2RGBDConversionUtils PUBLIC ${CMAKE_CURRENT_LIST_DIR})
 
-target_link_libraries(Ros2ConversionUtils PRIVATE YARP::YARP_os
+target_link_libraries(Ros2RGBDConversionUtils PRIVATE YARP::YARP_os
                                                      YARP::YARP_sig
                                                      rclcpp::rclcpp
                                                      sensor_msgs::sensor_msgs__rosidl_typesupport_cpp
                                                      std_msgs::std_msgs__rosidl_typesupport_c
                                                      YARP::YARP_dev)
 
-set_property(TARGET Ros2ConversionUtils PROPERTY FOLDER "Libraries/Msgs")
+set_property(TARGET Ros2RGBDConversionUtils PROPERTY FOLDER "Libraries/Msgs")
